Issue #556 Can't Update Tags for an Asset (#557)
For Issue #556, some methods did not include the specified `AssetPartitionId` when trying to resolve the tag. So modified the code such that it can resolve a unique tag without relying on the `AssetPartitionId`. Only when there are multiple tags with the same name will the `AssetPartitionId` be required.

We'll use the already existing `Resolve-AssetPartitionIdFromSafeguardSession` method, and also have it default to the macrocosm partition if not specified.